Move inprocess_dispatcher
